Functional VoIP prototype from client -> server
Changed the MicRecorder class to use WASAPI instead of WaveEvent. Added the StartListening() function to ClientApp, it is called immediately on binding to UDP. It creates a MicRecorder and uses it to capture microphone audio, then tries to send it to the server. Added a Pcm16Player class to Common, it will play whatever bytes you pass it using AddFrame(). In the UDP receive loop I just assume all UDP packets are audio (which is true for now) and just call AddFrame on each packets Payload. Removed some unneeded Server Notification.Invoke calls to avoid cluttering the console. Changed the ClientInterface code to use '127.0.0.1' instead of 'localhost', just in case I ever pass that variable to an IPAddress.Parse()