Fixed disconnection handling, added You: prefix to messages sent
Server disconnects are now handled properly by the client instead of throwing an exception. The bug where exiting the program after an error caused an exception has also been fixed. After a client sends a message, the cursor is now moved to the start of the line and 'You: ' is added before the message